+++ src/components/dashboard/content-row/content-element-wrapper/element-avatar.js
